run_python/requirements.txt
# Core dependencies
polygon-api-client>=1.12.0
boto3>=1.26.0
pandas>=2.0.0
pytz>=2023.3
python-dotenv>=1.0.0

# Async HTTP client
aiohttp>=3.8.0

# Additional dependencies for data processing
numpy>=1.24.0

# Logging and utilities
requests>=2.28.0